+++ utils_test.go
@@ -1,4 +1,4 @@
-package reflectil_test
+package structil_test
 
 import (
 	"errors"
@@ -7,7 +7,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 
-	. "github.com/goldeneggg/structil/reflectil"
+	. "github.com/goldeneggg/structil"
 )
 
 type tStruct struct {
@@ -191,6 +191,24 @@ func TestElemTypeOf(t *testing.T) {
 	}
 }
 
+func TestRecoverToError(t *testing.T) {
+	t.Parallel()
+
+	t.Run("recover to error", func(t *testing.T) {
+		defer func() {
+			if r := recover(); r != nil {
+				err := RecoverToError(r)
+				if err == nil {
+					t.Errorf("RecoverToError() did not return error. r: %v", r)
+					return
+				}
+				t.Logf("%v", err)
+			}
+		}()
+		panic("panic for test")
+	})
+}
+
 // benchmark tests
 
 func BenchmarkToI_String(b *testing.B) {
@@ -271,4 +289,20 @@ func benchmarkElemTypeOf(i interface{}, b *testing.B) {
 		typ = ElemTypeOf(i)
 		_ = typ
 	}
+}
+
+func BenchmarkRecoverToError(b *testing.B) {
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		recoverToErrorDemo()
+	}
+}
+
+func recoverToErrorDemo() {
+	defer func() {
+		if r := recover(); r != nil {
+			_ = RecoverToError(r)
+		}
+	}()
+	panic("panic for benchmark")
 }
